// apps/desktop/src/services/settingsCenter/schema.ts
export type ExtendedSettings = {
  startupPage: 'overview' | 'reader' | 'library' | 'last';
  openLastReaderBookOnStartup: boolean;
  restoreLastReaderPositionOnStartup: boolean;
  checkUnfinishedTasksOnStartup: boolean;
  refreshLibraryMetadataOnStartup: boolean;
  startupOverviewMode: 'auto' | 'welcome' | 'recent' | 'emptyGuide';
// ...
};